added first draft of curve LP tokens

+++ contracts/interfaces/IERC20WithDeciamls.sol
@@ -5,5 +5,5 @@ pragma solidity ^0.8.0;
 import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface IERC20WithDeciamls is IERC20 {
-    function decimals() external view returns (uint256);
+    function decimals() external view returns (uint8);
 }
